Drawing: organise drawing tool files Organise the drawing tool components and models into folders, in line with the file structure used in the rest of the code. One folder for each commponent plus its tests.
