+++ packages/lib-classifier/src/plugins/drawingTools/components/DrawingToolRoot/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrawingToolRoot'
