+++ packages/lib-classifier/src/plugins/drawingTools/components/Line/Line.spec.js
