+++ packages/lib-classifier/src/plugins/drawingTools/components/Point/Point.spec.js
