+++ packages/lib-classifier/src/plugins/drawingTools/components/Point/index.js
@@ -1,0 +1,1 @@
+export { default } from './Point'
