+++ packages/lib-classifier/src/plugins/drawingTools/models/marks/Line/index.js
@@ -1,0 +1,1 @@
+export { default } from './Line'
