+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/LineTool/LineTool.spec.js
