+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/LineTool/index.js
@@ -1,0 +1,1 @@
+export { default } from './LineTool'
