+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/PointTool/PointTool.spec.js
