+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/PointTool/index.js
@@ -1,0 +1,1 @@
+export { default } from './PointTool'
