+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/Tool/Tool.spec.js
