+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/Tool/index.js
@@ -1,0 +1,1 @@
+export { default } from './Tool'
